fix(admin): chat - address review comments
- auditClient: default to slog.Default() when a nil logger is passed
  (avoids a panic when logging); warn when only one of
  INTERNAL_CLIENT_CERT/INTERNAL_CLIENT_KEY is set so misconfiguration is
  observable; log CA file read/parse failures instead of ignoring them
- AdminCountMessages: repo now scans COUNT(*) into int64 (matches Postgres
  COUNT(*) width) and returns int64, dropping the service-layer cast;
  interface/mocks/tests updated accordingly
- tests: audit client nil-logger and partial-cert-config coverage

+++ backend/services/chat/query/repo/admin_test.go
@@ -109,6 +109,6 @@ func TestProjectionRepo_AdminCountMessages(t *testing.T) {
 
 	n, err := repo.AdminCountMessages(context.Background())
 	assert.NoError(t, err)
-	assert.Equal(t, 9, n)
+	assert.Equal(t, int64(9), n)
 	require.NoError(t, mock.ExpectationsWereMet())
 }
